Fixed a bug where error connecting to MySQL was not caught correctly
Also added functionality to attempt to flush crontab with -u <xbm_user>
when the flush is not being called by the user it should be installed in the crontab for.

Cleaned up some of the output of xbm-flush.php too

Also disabled PHP warnings in config.php by default for cleaner output.