Fixed a bad bug where any of the processes called by Xtrabackup Manager would never return. It would get stuck waiting for the process to end as a result of not reading from the output streams.
This only triggered when the process called had enough output to fill the stream buffer.

Also changed the innobackupex command used to specify --ibbackup= for the xtrabackup binary. Currently in xtrabackup-1.4 it seems this option is ignored. May decide to embed xtrabackup and scripts inside xtrabackup manager so that we can patch some issues like this.